Update AuthorService and TitleService
Added code to properly verifying the arguments given by the user.

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/librarysystem/service/AuthorService.java
@@ -26,6 +26,7 @@ public class AuthorService {
 
     @Transactional
     public Author createAuthor(String firstName, String lastName) {
+    	if (firstName == null || lastName == null || firstName.length() == 0 || lastName.length() == 0) throw new IllegalArgumentException("Please enter a valid name");
         Author author = new Author(firstName, lastName);
         authorRepository.save(author);
         return author;
@@ -33,13 +34,17 @@ public class AuthorService {
 
     @Transactional
     public Author createAuthor(String authorID, String firstName, String lastName) {
+    	if (authorID == null || authorID.length() == 0) throw new IllegalArgumentException("Please enter a valid authorID"); 
+    	if (firstName == null || lastName == null || firstName.length() == 0 || lastName.length() == 0) throw new IllegalArgumentException("Please enter a valid name");
         Author author = new Author(authorID, firstName, lastName);
         authorRepository.save(author);
         return author;
     }
+    
 
     @Transactional
     public Author getAuthorByAuthorID(String authorID) {
+    	if (authorID == null || authorID.length() == 0) throw new IllegalArgumentException("Please enter a valid authorID"); 
         Author author = authorRepository.findByAuthorID(authorID);
         if (author == null) throw new NullPointerException("Author not found");
         return author;
@@ -47,6 +52,7 @@ public class AuthorService {
 
     @Transactional
     public List<Author> getAuthorsByFirstName(String firstName) {
+    	if (firstName == null || firstName.length() == 0) throw new IllegalArgumentException("Please enter a valid first name");
         List<Author> authors = authorRepository.findByFirstName(firstName);
         if (authors.size() == 0) throw new NullPointerException("Authors not found");
         return authors;
@@ -54,6 +60,7 @@ public class AuthorService {
 
     @Transactional
     public List<Author> getAuthorsByLastName(String lastName) {
+    	if (lastName == null || lastName.length() == 0) throw new IllegalArgumentException("Please enter a valid last name");
         List<Author> authors = authorRepository.findByLastName(lastName);
         if (authors.size() == 0) throw new NullPointerException("Authors not found");
         return authors;
@@ -61,6 +68,7 @@ public class AuthorService {
 
     @Transactional
     public List<Author> getAuthorsByFirstNameAndLastName(String firstName, String lastName) {
+    	if (firstName == null || lastName == null || firstName.length() == 0 || lastName.length() == 0) throw new IllegalArgumentException("Please enter a valid name");
         List<Author> authors = authorRepository.findByFirstNameAndLastName(firstName, lastName);
         if (authors.size() == 0) throw new NullPointerException("Authors not found");
         return authors;
@@ -69,6 +77,7 @@ public class AuthorService {
 
     @Transactional
     public List<Author> getAuthorsByTitles(List<Title> titles) {
+    	if (titles == null || titles.size() == 0) throw new IllegalArgumentException("Please enter a valid list of titles");
         List<Author> authors = authorRepository.findByTitlesIn(titles);
         if (authors.size() == 0) throw new NullPointerException("Authors not found");
         return authors;
@@ -77,6 +86,7 @@ public class AuthorService {
 
     @Transactional
     public boolean deleteAuthorByAuthorID(String authorID) {
+    	if (authorID == null || authorID.length() == 0) throw new IllegalArgumentException("Please enter a valid authorID"); 
         Author author = authorRepository.findByAuthorID(authorID);
         if (author == null) throw new NullPointerException("Author not found");
         author.delete();
@@ -87,6 +97,7 @@ public class AuthorService {
 
     @Transactional
     public boolean deleteAuthorsByFirstName(String firstName) {
+    	if (firstName == null || firstName.length() == 0) throw new IllegalArgumentException("Please enter a valid first name");
         List<Author> authors = authorRepository.findByFirstName(firstName);
         if (authors.size() == 0) throw new NullPointerException("Authors not found");
         for (Author author : authors) {
@@ -99,6 +110,7 @@ public class AuthorService {
 
     @Transactional
     public boolean deleteAuthorsByLastName(String lastName) {
+    	if (lastName == null || lastName.length() == 0) throw new IllegalArgumentException("Please enter a valid last name");
         List<Author> authors = authorRepository.findByLastName(lastName);
         if (authors.size() == 0) throw new NullPointerException("Authors not found");
         for (Author author : authors) {
@@ -111,6 +123,7 @@ public class AuthorService {
 
     @Transactional
     public boolean deleteAuthorsByFirstNameAndLastName(String firstName, String lastName) {
+    	if (firstName == null || lastName == null || firstName.length() == 0 || lastName.length() == 0) throw new IllegalArgumentException("Please enter a valid name");
         List<Author> authors = authorRepository.findByFirstNameAndLastName(firstName, lastName);
         if (authors.size() == 0) throw new NullPointerException("Authors not found");
         for (Author author : authors) {
@@ -137,21 +150,25 @@ public class AuthorService {
 
     @Transactional
     public boolean isAuthorsExistsByFirstName(String firstName) {
+    	if (firstName == null || firstName.length() == 0) throw new IllegalArgumentException("Please enter a valid first name");
         return (authorRepository.existsByFirstName(firstName));
     }
 
     @Transactional
     public boolean isAuthorsExistsByLastName(String lastName) {
+    	if (lastName == null || lastName.length() == 0) throw new IllegalArgumentException("Please enter a valid last name");
         return (authorRepository.existsByLastName(lastName));
     }
 
     @Transactional
     public boolean isAuthorsExistsByFirstNameAndLastName(String firstName, String lastName) {
+    	if (firstName == null || lastName == null || firstName.length() == 0 || lastName.length() == 0) throw new IllegalArgumentException("Please enter a valid name");
         return (authorRepository.existsByFirstNameAndLastName(firstName, lastName));
     }
 
     @Transactional
     public boolean isAuthorsExistsByAuthorID(String authorID) {
+    	if (authorID == null || authorID.length() == 0) throw new IllegalArgumentException("Please enter a valid authorID"); 
         return (authorRepository.existsByAuthorID(authorID));
     }
