+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/librarysystem/service/TitleService.java
@@ -27,6 +27,9 @@ public class TitleService {
 
     @Transactional
     public Title createTitle(String name, String pubDate, Author... author) {
+    	if (name == null || name.length() == 0) throw new IllegalArgumentException("Please enter a valid title name");
+    	if (pubDate == null || pubDate.length() == 0) throw new IllegalArgumentException("Please enter a valid publish date");
+    	if (author == null || author.length == 0) throw new IllegalArgumentException("Please enter a list of valid authors"); 
         Title title = new Title(name,pubDate,author);
         titleRepository.save(title);
         return title;
@@ -34,6 +37,7 @@ public class TitleService {
 
     @Transactional
     public Title getTitleByTitleID(String titleID) {
+    	if (titleID == null || titleID.length() == 0) throw new IllegalArgumentException("Please enter a valid titleID"); 
         Title title = titleRepository.findByTitleID(titleID);
         if (title == null) throw new NullPointerException("Title not found");
         return title;
@@ -41,6 +45,7 @@ public class TitleService {
 
     @Transactional
     public List<Title> getTitlesByAuthorID(Author authorID) {
+    	if (authorID == null) throw new IllegalArgumentException("Please enter a valid author"); 
         List<Title> titles = titleRepository.findByAuthor(authorID);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         return titles;
@@ -48,6 +53,7 @@ public class TitleService {
 
     @Transactional
     public List<Title> getTitlesByAuthorIDs(List<Author> authorID) {
+    	if (authorID == null || authorID.size() == 0) throw new IllegalArgumentException("Please enter a list of valid authors"); 
         List<Title> titles = titleRepository.findByAuthorIn(authorID);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         return titles;
@@ -55,6 +61,7 @@ public class TitleService {
 
     @Transactional
     public Title getTitlesByItemBarcode(Item itemBarcode) {
+    	if (itemBarcode == null) throw new IllegalArgumentException("Please enter a valid item");
         Title title = titleRepository.findByItem(itemBarcode);
         if (title == null) throw new NullPointerException("Title not found");
         return title;
@@ -62,6 +69,7 @@ public class TitleService {
 
     @Transactional
     public List<Title> getTitlesByItemBarcodes(List<Item> itemBarcode) {
+    	if (itemBarcode == null || itemBarcode.size() == 0) throw new IllegalArgumentException("Please enter a valid list of items");
         List<Title> titles = titleRepository.findByItemIn(itemBarcode);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         return titles;
@@ -69,6 +77,7 @@ public class TitleService {
 
     @Transactional
     public List<Title> getTitlesByName(String name) {
+    	if (name == null || name.length() == 0) throw new IllegalArgumentException("Please enter a valid title name");
         List<Title> titles = titleRepository.findByName(name);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         return titles;
@@ -76,6 +85,7 @@ public class TitleService {
 
     @Transactional
     public List<Title> getTitlesByPubDate(String pubDate) {
+    	if (pubDate == null || pubDate.length() == 0) throw new IllegalArgumentException("Please enter a valid publish date");
         List<Title> titles = titleRepository.findByPubDate(pubDate);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         return titles;
@@ -83,6 +93,8 @@ public class TitleService {
 
     @Transactional
     public Title getTitleByNameAndPubDate(String name, String pubDate) {
+    	if (name == null || name.length() == 0) throw new IllegalArgumentException("Please enter a valid title name");
+    	if (pubDate == null || pubDate.length() == 0) throw new IllegalArgumentException("Please enter a valid publish date");
         Title title = titleRepository.findByNameAndPubDate(name, pubDate);
         if (title == null) throw new NullPointerException("Title not found");
         return title;
@@ -90,6 +102,7 @@ public class TitleService {
 
     @Transactional
     public boolean deleteTitleByTitleID(String titleID) {
+    	if (titleID == null || titleID.length() == 0) throw new IllegalArgumentException("Please enter a valid titleID"); 
         Title title = titleRepository.findByTitleID(titleID);
         if (title == null) throw new NullPointerException("Title not found");
         title.delete();
@@ -100,6 +113,7 @@ public class TitleService {
 
     @Transactional
     public boolean deleteTitlesByAuthorID(Author authorID) {
+    	if (authorID == null) throw new IllegalArgumentException("Please enter a valid authorID"); 
         List<Title> titles = titleRepository.findByAuthor(authorID);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         for (Title title : titles) {
@@ -112,6 +126,7 @@ public class TitleService {
 
     @Transactional
     public boolean deleteTitlesByAuthorIDs(List<Author> authorID) {
+    	if (authorID == null || authorID.size() == 0) throw new IllegalArgumentException("Please enter a list of valid authors"); 
         List<Title> titles = titleRepository.findByAuthorIn(authorID);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         for (Title title : titles) {
@@ -125,6 +140,7 @@ public class TitleService {
 
     @Transactional
     public boolean deleteTitlesByItemBarcode(Item itemBarcode) {
+    	if (itemBarcode == null) throw new IllegalArgumentException("Please enter a valid item"); 
         Title title = titleRepository.findByItem(itemBarcode);
         if (title == null) throw new NullPointerException("Title not found");
         title.delete();
@@ -136,6 +152,7 @@ public class TitleService {
 
     @Transactional
     public boolean deleteTitlesByItemBarcodes(List<Item> itemBarcode) {
+    	if (itemBarcode == null || itemBarcode.size() == 0) throw new IllegalArgumentException("Please enter a valid list of items");
         List<Title> titles = titleRepository.findByItemIn(itemBarcode);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         for (Title title : titles) {
@@ -149,6 +166,7 @@ public class TitleService {
 
     @Transactional
     public boolean deleteTitlesByName(String name) {
+    	if (name == null || name.length() == 0) throw new IllegalArgumentException("Please enter a valid title name");
         List<Title> titles = titleRepository.findByName(name);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         for (Title title : titles) {
@@ -161,6 +179,7 @@ public class TitleService {
 
     @Transactional
     public boolean deleteTitlesByPubDate(String pubDate) {
+    	if (pubDate == null || pubDate.length() == 0) throw new IllegalArgumentException("Please enter a valid publish date");
         List<Title> titles = titleRepository.findByPubDate(pubDate);
         if (titles.size() == 0) throw new NullPointerException("Titles not found");
         for (Title title : titles) {
@@ -173,7 +192,9 @@ public class TitleService {
 
     @Transactional
     public boolean deleteTitleByNameAndPubDate(String name, String pubDate) {
-        Title title = titleRepository.findByNameAndPubDate( name, pubDate);
+    	if (name == null || name.length() == 0) throw new IllegalArgumentException("Please enter a valid title name");
+    	if (pubDate == null || pubDate.length() == 0) throw new IllegalArgumentException("Please enter a valid publish date");
+        Title title = titleRepository.findByNameAndPubDate(name, pubDate);
         if (title == null) throw new NullPointerException("Title not found");
         title.delete();
         titleRepository.delete(title);
@@ -183,16 +204,20 @@ public class TitleService {
 
     @Transactional
     public boolean isTitleExistsByTitleID(String titleID) {
+    	if (titleID == null || titleID.length() == 0) throw new IllegalArgumentException("Please enter a valid titleID"); 
         return (titleRepository.existsByTitleID(titleID));
     }
 
     @Transactional
     public boolean isTitleExistsByItem(Item itemBarcode) {
+    	if (itemBarcode == null) throw new IllegalArgumentException("Please enter a valid item"); 
         return (titleRepository.existsByItem(itemBarcode));
     }
 
     @Transactional
     public boolean isTitleExistsByNameAndPubDate(String name, String pubDate) {
+    	if (name == null || name.length() == 0) throw new IllegalArgumentException("Please enter a valid title name");
+    	if (pubDate == null || pubDate.length() == 0) throw new IllegalArgumentException("Please enter a valid publish date");
         return (titleRepository.existsByNameAndPubDate(name, pubDate));
     }
